+++ demo/api/src/db/fixtures/generators/blocks/media/dam-video-block-fixture.service.ts
@@ -1,5 +1,4 @@
-import { ExtractBlockInputFactoryProps } from "@comet/blocks-api";
-import { DamVideoBlock } from "@comet/cms-api";
+import { DamVideoBlock, ExtractBlockInputFactoryProps } from "@comet/cms-api";
 import { faker } from "@faker-js/faker";
 import { Injectable } from "@nestjs/common";