+++ demo/api/src/db/fixtures/generators/blocks/media/media-block.fixture.service.ts
@@ -1,4 +1,4 @@
-import { ExtractBlockInputFactoryProps } from "@comet/blocks-api";
+import { ExtractBlockInputFactoryProps } from "@comet/cms-api";
 import { faker } from "@faker-js/faker";
 import { Injectable } from "@nestjs/common";
 import { MediaBlock } from "@src/common/blocks/media.block";