+++ packages/admin/admin/src/form/fields/SwitchField.tsx
@@ -1,14 +1,11 @@
-<<<<<<< HEAD
 import { FormControlLabel, type FormControlLabelProps } from "@mui/material";
-=======
-import { FormControlLabel, FormControlLabelProps } from "@mui/material";
-import { ReactNode } from "react";
->>>>>>> main
+import { type ReactNode } from "react";
 
 import { Field, type FieldProps } from "../Field";
 import { FinalFormSwitch, type FinalFormSwitchProps } from "../Switch";
 
-export interface SwitchFieldProps extends FieldProps<string, HTMLInputElement> {
+export interface SwitchFieldProps extends Omit<FieldProps<string, HTMLInputElement>, "label"> {
+    name: string;
     fieldLabel?: ReactNode;
     label?: ReactNode | ((checked?: boolean) => ReactNode);
     componentsProps?: {