+++ packages/admin/cms-admin/src/blocks/factories/createColumnsBlock.tsx
@@ -407,13 +407,8 @@ export function createColumnsBlock<T extends BlockInterface>(
                                                                 size="large"
                                                                 startIcon={<Add />}
                                                             >
-<<<<<<< HEAD:packages/admin/cms-admin/src/blocks/factories/createColumnsBlock.tsx
-                                                                <FormattedMessage id="comet.blocks.columns.addColumn" defaultMessage="Add column" />
-                                                            </BlockAdminComponentButton>
-=======
                                                                 <FormattedMessage id="comet.blocks.columns.addItem" defaultMessage="Add item" />
-                                                            </AdminComponentButton>
->>>>>>> main:packages/admin/blocks-admin/src/blocks/factories/createColumnsBlock.tsx
+                                                            </BlockAdminComponentButton>
                                                         </AdminComponentStickyFooter>
                                                     )}
                                                 </SelectPreviewComponent>