+++ storybook/.storybook/preview.tsx
@@ -1,12 +1,9 @@
 import "@fontsource-variable/roboto-flex/full.css";
 
-<<<<<<< HEAD
-import { createCometTheme, MainContent, MuiThemeProvider } from "@comet/admin";
-=======
-import { DataGridPanel, MainContent, MuiThemeProvider } from "@comet/admin";
->>>>>>> main
+import { createCometTheme, DataGridPanel, MainContent, MuiThemeProvider } from "@comet/admin";
 import { DateFnsLocaleProvider } from "@comet/admin-date-time";
 import { createTheme as createMuiTheme, GlobalStyles } from "@mui/material";
+import type {} from "@mui/x-data-grid/themeAugmentation";
 import type { Preview } from "@storybook/react";
 import { type Locale as DateFnsLocale } from "date-fns";
 import { de as deLocale, enUS as enLocale } from "date-fns/locale";
@@ -83,8 +80,9 @@ const preview: Preview = {
                           components: {
                               MuiDataGrid: {
                                   defaultProps: {
-                                      components: {
-                                          Panel: DataGridPanel,
+                                      slots: {
+                                          // @ts-expect-error @jamesricky fix this please
+                                          panel: DataGridPanel,
                                       },
                                   },
                               },