+++ packages/site/cms-site/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Patch Changes
 
--   75fb1d0d4: Fix block preview not rendering before user interaction
+- 75fb1d0d4: Fix block preview not rendering before user interaction
 
 ## 7.14.0